Improve the connect/disconnect flow
Several changes have been made in the main process and in shared in
order for the connection state to be synchronized between renderers.
This commit implements those changes.

This is a temporary solution, and improved UX must be worked on in the
future, with regards to overwrite functionality when the terminal tries
to open a port. Now there's a toggle for it, but this must be reviewed.