Add option to open port with current settings
In cases where a different app has claimed the selected port with
different settings, we want to enable the action of opening with those
settings that are already in use.

Additionally, this commit gives the user a bit more knowledge about the
state, by providing the selected settings in the current app, and also
the options which are currently in use.