Extract Baud Rate setting to it's own file
The SerialSettings file is quite overwhelming, and it could benefit some
decouplings in order to make it easier to comprehend.

This commit introduces the NumberInputWithDropdown component, which is
implemented in pc-nrfconnect-shared v121, and makes the custom baud rate
more integrated into the baudrate selection.

The previous implementation was a good starting point, but it felt like
the custom field and the dropdown menu didn't belong together, but
functionally they do. This is the motivation to merge them together.

NB! The next commit needs to update shared to v121.