Allow user to set the scrollback value (#42)
The scrollback value determines how many lines will be kept in the
XTerms buffer, and determines how many lines you are able to
scroll back in history.

* Do 'live' update of scrollback

XTerm is able to adjust the buffer size 'on-the-fly'. Meaning that while
a device is connected, there is no problem increasing or decreasing the
scrollback value.

Note that if there is data in the buffer, and the scrollback is
decreased, then the data in the buffer that goes beyond the new
scrollback limit, will be removed.