Add .history file
The .history file should record all the commands that are sent by the
user of Terminal. It has a HISTORY_MAX_SIZE, which says ~ how many lines
can be recorded.

When HISTORY_MAX_SIZE number of lines, plus a threshold, is reached,
then we truncate the content of the file, but keep exactly
HISTORY_MAX_SIZE number of lines in the file.