Sync read/write between terminal instances
This commit introduces a History Buffer, which utilizes a closure
in order to store the whole history content in-memory. This is first
and foremost to improve performance.

When testing, I found on my machine, that reading took 2x longer than
writing, and processing the data after reading was about 1-to-1 with
writing the data.

Assuming every read-write iteration uses these steps read - process -
write. It takes ~300% longer to go through the read-write cycle, instead
of just writing what is already in buffer.

The down-side to storing the content in buffer, is that we cannot be
sure that the content in the file is 100% in-sync with what is
in-memory.

To mitigate this issue, we assume that in order to write to the history
file, the app must be in-focus. Hence, after not being in focus, we must
always update the content in-memory, with what is in the file.