Add iterative scroll back after typing
As a user of Serial Terminal I want to find previously used
commands by clicking arrow-up and arrow-down to browse through
the history.

If I have started typing, and then click arrow-up, I want to
find the commands that start with the same text as I've already typed
into the input field.

Note: This feature is only relevant for Line Mode, as it should be
implemented by the shell, running on the device, when shell mode is
used.