Add Terminal Settings Pane
Added the HistoryFile settings, as it doesn't seem to belong in
the side panel, as it will be more like a one-time change, rather
than a per-session change of settings.

This commit also makes the general history file settings interaction
even more minimal, as the progress bar is removed, and rather just
utilizes a simple percentage.