Move more permanent settings into separate Pane
Where it is more suitable, we should put settings into a different Pane.
The Side Panel, where we currently have all the settings, should only
contain settings that are relevant to customize for any session.

In contrast, the settings that should preferably only change once,
should be put into the Terminal Settings Pane.