index for backward search must be updated
When the history file is full, the history buffer is also trimmed,
but only after the new entry is pushed to the list and the
currentLineIndex is updated.

In order for the index to be properly updated, an additional update
is added to the trimming function. Else the index would be one-off, and
it would require the user to click arrow up twice to see the last entry.